requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anthropic==0.52.1
anyio==4.9.0
argcomplete==3.6.2
boto3==1.38.27
botocore==1.38.27
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
fasta2a==0.2.12
fastapi==0.115.12
fastapi-cli==0.0.7
fastavro==1.11.1
filelock==3.18.0
fpdf==1.7.2
fsspec==2025.5.1
google-auth==2.40.2
google-genai==1.18.0
greenlet==3.2.2
griffe==1.7.3
groq==0.26.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
logfire-api==3.16.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.2
mdurl==0.1.2
mistralai==1.8.1
openai==1.82.1
opentelemetry-api==1.33.1
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-ai==0.2.12
pydantic-ai-slim==0.2.12
pydantic-evals==0.2.12
pydantic-graph==0.2.12
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
s3transfer==0.13.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
sse-starlette==2.3.6
starlette==0.46.2
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.0.20250602
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
zipp==3.22.0
